Match test function signatures

+++ include/tcp.h
@@ -18,7 +18,7 @@ typedef struct {
   struct sockaddr_in address;
 } tcp_server;
 
-server_status_e bind_tcp_port(tcp_server *server, uint16_t port);
+server_status_e bind_tcp_port(tcp_server *server, int port);
 int accept_client(int server_fd);
 
 #endif // !TCP_H
